Add files via upload
Project Title: Exploratory Data Analysis on the Titanic Dataset

Description:

In this project, I conducted a comprehensive Exploratory Data Analysis (EDA) on the famous Titanic dataset, employing Python and its data science libraries. The objective was to uncover patterns and insights that shed light on the factors influencing passenger survival during the ill-fated maiden voyage of the RMS Titanic.